scrobbler: Retry scrobble for unsuccessful 'stop' scrobbles
Using the /sync/history API endpoint of trakt, scrobbles which could
not be completed(most likely due to network failure) are saved in local
cache, which is cleaned periodically. Only stop scrobbles which can
affect user watched history are used, so play and pause scrobbles are
still discarded on failure.