monitor: Never modify media_info dict
In case of multiple episodes, the parse_status used to put the updated
episode number directly inside the media_info dict from get_media_info,
which is actually cached. And since dicts are mutable, this meant that
for all subsequent calls to get_media_info, it would get the modified
dict with a single episode instead of a list of episode numbers.