plex: Add "scrobble_user" to limit scrobbling to a single Plex user
Extending this to make it work with multiple plex users opnes a very big
can of worms. For instance, how to handle 2 (or more) users in the
whitelist who are playing media at the same time? What should be marked
as "Currently watching" in Trakt. Should there be a priority of users?

For now, only one plex user can be whitelisted.

Closes #80