Add preview and fast_pause features
The preview feature gets activated when starting a file at >80% progress
and doesn't scrobble anything for 60 seconds. This is to prevent
redundant "watched" scrobbles.

Fast pause is mainly for MPV (or any future monitor that supports live-
updates). If the user does a series of "pause->play->pause->play..."
actions in quick succession (possibly due to the "next-frame" command),
this would cause each action to get scrobbled. Now once the monitor sees
a transition "paused->playing->paused" within 1 second, it activates
"fast_pause" state, where further plays/pauses are ignored. To exit this
state, there should be no "pause" for 5 seconds.

This code immensely complicates the state transition logic for the
monitor due to the addition of timers. Lots of testing required.