MPC-HC: Fix unhandled KeyError in variable extraction
For some reason, when the file starts, MPC-HC creates the variables
page, but doesn't actually populate it until a few seconds(?) later. But
our scrobbler assumes that if the webpage is there, the variables are
also present. Since this wasn't the case, we were crashing. The fix is
to simply add another check that the necessary variables are present,
since they will be captured during the next poll.

Fixes #115