Fix poetry install failing for urlmatch
This error is happening because poetry cannot identify legacy versions
in sdist names, and it tries to install a macos-specific package that
doesn't contain a setup.py or setup.cfg file. The workaround is to
blacklist the macos package file by corrupting its hash value in the
lock file, so poetry will be forced to install from the other file.

Needs poetry v1.2 or higher to actually respect the missing hash. Lower
versions simply ignore hash mismatch, and install the package anyway.