logger: Only log 'property unavailable' from mpv once
The log message is generated when mpv is open but nothing is playing.
The mpvpoll thread keeps pinging the player regardless, and some
properties are unavailable at this time. We could try to disable
mpvpoll when nothing is playing, but detecting this might be more
trouble than it is worth.