file_info: only use the actual path for guessit, skipping other parts
Guessit expects only the filename, and it was mostly by chance that it
accepted/correctly parsed URLs with domain names. But it fails when the
host is an IP Address, so better to fix it permanently by only using
the path portion of the URL.

Will have to see if the params portion is required for some sites, in
which case more rethinking is required. Particularly, a better config
format that can differentiate between local and remote paths, and what
portions of remote path to use.

Fixes #125