file_info: Split whitelist at module load time
Instead of relying on BadMatchPattern every single time when trying to
whitelist urls, now we split the whitelist into local and remote
whitelisted paths directly when the module is loaded. This makes the
whitelist function itself slightly more efficient since it doesn't have
to rely on exception handling of BadMatchPattern for non-urls.