notifier: Allow toggling updatable_notifs on Linux
A recent update to KDE Plasma (v5.22) has broken the behaviour of
replaces_id in the notification daemon.
See https://bugs.kde.org/show_bug.cgi?id=440051

We need to add more bookkeeping to track whether the previous
notification has closed, since we can only use the replaces_id for a
live notification. It seems to be more trouble than it is worth, better
to kill this feature completely.