notifier: Use desktop-notifier on all platforms
This is a WIP with lots of hacks

One interesting thing is the pythoncom clash with winrt.
Somehow, when using windows_toasts package instead of desktop-notifier,
we don't need to instantiate the notifier object. Importing it before
pythoncom is enough. Should raise a bug report- but not sure where :P

Closes #259