monitor: Improve jump detection
Earlier, we used to simply calc the delta of file progress % between the
prev and cur states. If too much delta from last update, send scrobble.
But for small files (like, 2 minute duration), with a polling monitor
running every 10 seconds, we would hit this condition after each check.
Because, the progress delta in 10s is a significant chunk of the total
time. This would result in spurious "start" scrobbles every 10s.

Now we also estimate the "expected" progress between the two states,
based on their "updated_at" times as a % of total duration. Now, if
this "expected" progress is too far from the measured progress, we send
the scrobble. This assumes that the file is being played at 1x speed.